:recycle: model layer 삭제
